requirements.txt
requests
pycryptodome
fake-useragent
python-dotenv